+++ packages/gafl-webapp-service/src/handlers/__tests__/analytics-handler.spec.js
@@ -2,7 +2,9 @@ import { ANALYTICS } from '../../constants.js'
 import analyticsHandler from '../analytics-handler.js'
 import { addLanguageCodeToUri } from '../../processors/uri-helper.js'
 
-jest.mock('../../processors/uri-helper.js')
+jest.mock('../../processors/uri-helper.js', () => ({
+  addLanguageCodeToUri: jest.fn((_request, uri) => uri)
+}))
 
 jest.mock('../../constants', () => ({
   ANALYTICS: {
@@ -17,44 +19,71 @@ describe('The analytics handler', () => {
     jest.clearAllMocks()
   })
 
-  it('redirects to /buy', async () => {
-    const payload = {
-      analyticsResponse: 'accept'
-    }
-    const request = generateRequestMock(payload)
+  it.each([
+    ['https://localhost:3000', 'https://localhost:3000/buy/name', 'localhost:3000', '/buy/name'],
+    ['https://localhost:1234', 'https://localhost:1234/example/test/redirect', 'localhost:1234', '/example/test/redirect'],
+    [
+      'https://testserver-example-fish',
+      'https://testserver-example-fish/buy/renew/identify',
+      'testserver-example-fish',
+      '/buy/renew/identify'
+    ]
+  ])('redirects to correct page if the HTTP_REFERER host matches the host of the current page', async (origin, referer, host, redirect) => {
+    const payload = { analyticsResponse: 'accept' }
+    const headers = { origin, referer }
+    const request = generateRequestMock(payload, 'analytics', headers, host)
     const responseToolkit = generateResponseToolkitMock()
+
     await analyticsHandler(request, responseToolkit)
-    expect(responseToolkit.redirect).toHaveBeenCalledWith(addLanguageCodeToUri(request, '/buy'))
+
+    expect(responseToolkit.redirect).toHaveBeenCalledWith(redirect)
   })
 
-  it('get calls addLanguageCodeToUri with request and /buy', async () => {
-    const buyUri = '/buy'
-    const payload = {
-      analyticsResponse: 'accept'
-    }
+  it.each([
+    ['https://localhost:3000', 'https://localhost:3000/buy/name', 'localhost:3047'],
+    ['https://localhost:1234', 'https://localhost:1234/example/test/redirect', 'notsamehost:1234'],
+    ['https://testserver-example-fish', 'https://testserver-example-fish/buy/renew/identify', 'hdfhdskfhs-ghj-vgjh'],
+    [
+      'https://get-a-fishing-licence.gov.uk/buy/licence-for',
+      'https://some.shady.host/not/where/we/should/go',
+      'get-a-fishing-licence.gov.uk'
+    ]
+  ])('redirects to /buy if the HTTP_REFERER host does not match the host of the current page', async (origin, referer, host) => {
+    const payload = { analyticsResponse: 'accept' }
+    const headers = { origin, referer }
+    const request = generateRequestMock(payload, 'analytics', headers, host)
     const responseToolkit = generateResponseToolkitMock()
-    const request = generateRequestMock(payload)
 
     await analyticsHandler(request, responseToolkit)
 
-    expect(addLanguageCodeToUri).toHaveBeenCalledWith(request, buyUri)
+    expect(responseToolkit.redirect).toHaveBeenCalledWith('/buy')
+  })
+
+  it('calls addLanguageCodeToUri with request and /buy', async () => {
+    const headers = { origin: 'https://localhost:3000', referer: 'https://localhost:3000/buy' }
+    const redirect = '/buy'
+    const payload = { analyticsResponse: 'accept' }
+    const responseToolkit = generateResponseToolkitMock()
+    const request = generateRequestMock(payload, 'analytics', headers)
+
+    await analyticsHandler(request, responseToolkit)
+
+    expect(addLanguageCodeToUri).toHaveBeenCalledWith(request, redirect)
   })
 
   it('selected not true and response is accept sets selected to true and acceptTracking to true', async () => {
-    const payload = {
-      analyticsResponse: 'accept'
-    }
+    const payload = { analyticsResponse: 'accept' }
+    const request = generateRequestMock(payload, 'analytics')
 
-    const request = generateRequestMock(payload)
     await analyticsHandler(request, generateResponseToolkitMock())
+
     expect(mockAnalyticsSet).toHaveBeenCalledWith(expect.objectContaining({ [ANALYTICS.selected]: true, [ANALYTICS.acceptTracking]: true }))
   })
 
   it('selected not true and response is reject sets selected to true and acceptTracking to false', async () => {
-    const payload = {
-      analyticsResponse: 'reject'
-    }
-    const request = generateRequestMock(payload)
+    const payload = { analyticsResponse: 'reject' }
+    const request = generateRequestMock(payload, 'analytics')
+
     await analyticsHandler(request, generateResponseToolkitMock())
 
     expect(mockAnalyticsSet).toHaveBeenCalledWith(
@@ -63,11 +92,9 @@ describe('The analytics handler', () => {
   })
 
   it('selected is true sets seenMessage to true', async () => {
-    const analytics = {
-      [ANALYTICS.selected]: true
-    }
-
+    const analytics = { [ANALYTICS.selected]: true }
     const request = generateRequestMock('payload', analytics)
+
     await analyticsHandler(request, generateResponseToolkitMock())
 
     expect(mockAnalyticsSet).toHaveBeenCalledWith(expect.objectContaining({ [ANALYTICS.seenMessage]: true }))
@@ -75,9 +102,15 @@ describe('The analytics handler', () => {
 
   const mockAnalyticsSet = jest.fn()
 
-  const generateRequestMock = (payload, analytics = {}) => ({
+  const generateRequestMock = (
+    payload,
+    analytics,
+    headers = { origin: 'http://host/current/page', referer: 'http://host/previous/page' },
+    host = 'host'
+  ) => ({
     payload,
     url: {
+      host,
       search: ''
     },
     cache: jest.fn(() => ({
@@ -87,7 +120,8 @@ describe('The analytics handler', () => {
           set: mockAnalyticsSet
         }
       }
-    }))
+    })),
+    headers
   })
 
   const generateResponseToolkitMock = () => ({