+++ packages/gafl-webapp-service/src/handlers/page-handler.js
@@ -1,12 +1,28 @@
 import db from 'debug'
 import { CacheError } from '../session-cache/cache-manager.js'
 import { PAGE_STATE, ANALYTICS } from '../constants.js'
-import { CONTROLLER, PROCESS_ANALYTICS_PREFERENCES } from '../uri.js'
+import {
+  AGREED,
+  CONTROLLER,
+  LICENCE_DETAILS,
+  ORDER_COMPLETE,
+  PAYMENT_CANCELLED,
+  PAYMENT_FAILED,
+  PROCESS_ANALYTICS_PREFERENCES
+} from '../uri.js'
 import GetDataRedirect from './get-data-redirect.js'
 import journeyDefinition from '../routes/journey-definition.js'
 import { addLanguageCodeToUri } from '../processors/uri-helper.js'
 
 const debug = db('webapp:page-handler')
+const pagesToOmitAnalyticsBanner = [AGREED.uri, LICENCE_DETAILS.uri, ORDER_COMPLETE.uri, PAYMENT_CANCELLED.uri, PAYMENT_FAILED.uri]
+
+const displayAnalytics = request => {
+  if (pagesToOmitAnalyticsBanner.includes(request.path)) {
+    return false
+  }
+  return request.path.startsWith('/buy')
+}
 
 /**
  * Flattens the error structure from joi for use in the templates
@@ -115,6 +131,8 @@ export default (path, view, completion, getData) => ({
     pageData.analyticsMessageDisplayed = analytics[ANALYTICS.seenMessage]
     pageData.analyticsSelected = analytics[ANALYTICS.selected]
     pageData.acceptedTracking = analytics[ANALYTICS.acceptTracking]
+
+    pageData.displayAnalytics = displayAnalytics(request)
 
     return h.view(view, pageData)
   },